requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0
sentry-sdk==1.38.0
apscheduler==3.10.4
ruff==0.1.6
mypy==1.7.1 